pyproject.toml
[tool.poetry]
name = "tortoise-gis"
version = "0.1.2"
description = "Geometrical and Geographical support for Tortoise ORM."
authors = ["Eduardo Rezende <eduardorbr7@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/revensky/tortoise-gis"
repository = "https://github.com/revensky/tortoise-gis.git"
keywords = ["Tortoise", "GIS"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
packages = [{ include = "tortoise_gis" }]

[tool.poetry.dependencies]
python = "^3.8"
tortoise-orm = "^0.19.3"
shapely = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
black = "^23.3.0"
asyncpg = "^0.27.0"
python-dotenv = "^1.0.0"
rstcheck = "^6.1.2"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry>=1.5.0"]
build-backend = "poetry.masonry.api"
